smart-cv/data-processor/requirement.txt
grpcio-tools
python-dotenv
pytesseract
pdf2image
confluent-kafka[avro,json,protobuf]
minio
elasticsearch
scikit-learn
sentence-transformers
transformers
datasets
seqeval
torch
accelerate